KW-624/refactor/pass count 분리| pass -> stats
