--- src/main/java/com/doubleo/passservice/domain/pass/service/PassCountService.java
@@ -1,9 +1,0 @@
-package com.doubleo.passservice.domain.pass.service;
-
-import com.doubleo.passservice.domain.pass.dto.request.PassCountInfoRequest;
-import com.doubleo.passservice.domain.pass.dto.response.PassCountInfoResponse;
-import java.util.List;
-
-public interface PassCountService {
-    List<PassCountInfoResponse> getPassCount(PassCountInfoRequest request);
-}
